Assignment-3/pyproject.toml
[project]
name = "conceptguru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "usman-at-whizz-bridge", email = "usman.amir@whizzbridge.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]

[project.scripts]
conceptguru = "conceptguru:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
